# [307] Add support for histograms and distributions without unit conversion