--- a/exporter_benchmark_test.go
+++ b/exporter_benchmark_test.go
@@ -86,13 +86,13 @@ func BenchmarkExporterListener(b *testing.B) {
 			GMetricName: "gauge",
 			GValue:      10,
 		},
-		&event.TimerEvent{ // simple timer
-			TMetricName: "timer",
-			TValue:      200,
+		&event.ObserverEvent{ // simple timer
+			OMetricName: "timer",
+			OValue:      200,
 		},
-		&event.TimerEvent{ // simple histogram
-			TMetricName: "histogram.test",
-			TValue:      200,
+		&event.ObserverEvent{ // simple histogram
+			OMetricName: "histogram.test",
+			OValue:      200,
 		},
 		&event.CounterEvent{ // simple_tags
 			CMetricName: "simple_tags",