--- a/pkg/line/line.go
+++ b/pkg/line/line.go
@@ -43,16 +43,16 @@ func buildEvent(statType, metric string, value float64, relative bool, labels ma
 			GLabels:     labels,
 		}, nil
 	case "ms":
-		return &event.TimerEvent{
-			TMetricName: metric,
-			TValue:      float64(value) / 1000, // prometheus presumes seconds, statsd millisecond
-			TLabels:     labels,
+		return &event.ObserverEvent{
+			OMetricName: metric,
+			OValue:      float64(value) / 1000, // prometheus presumes seconds, statsd millisecond
+			OLabels:     labels,
 		}, nil
 	case "h", "d":
-		return &event.TimerEvent{
-			TMetricName: metric,
-			TValue:      float64(value),
-			TLabels:     labels,
+		return &event.ObserverEvent{
+			OMetricName: metric,
+			OValue:      float64(value),
+			OLabels:     labels,
 		}, nil
 	case "s":
 		return nil, fmt.Errorf("no support for StatsD sets")