--- a/pkg/mapper/metric_type.go
+++ b/pkg/mapper/metric_type.go
@@ -21,6 +21,7 @@ const (
 	MetricTypeCounter  MetricType = "counter"
 	MetricTypeGauge    MetricType = "gauge"
 	MetricTypeObserver MetricType = "observer"
+	MetricTypeTimer    MetricType = "timer" // DEPRECATED
 )
 
 func (m *MetricType) UnmarshalYAML(unmarshal func(interface{}) error) error {
@@ -35,6 +36,8 @@ func (m *MetricType) UnmarshalYAML(unmarshal func(interface{}) error) error {
 	case MetricTypeGauge:
 		*m = MetricTypeGauge
 	case MetricTypeObserver:
+		*m = MetricTypeObserver
+	case MetricTypeTimer:
 		*m = MetricTypeObserver
 	default:
 		return fmt.Errorf("invalid metric type '%s'", v)