This fixes #307 by moving millisecond-->second conversion to the buildEvent function for `ms` statTypes and also adds a case for `h` and `d` statTypes without the conversion. This effectively makes the TimerEvent value in seconds/unitless. Since this is no longer just limited to timers, should the TimerEvent be renamed to DistributionEvent or something similar? 

Test cases were also added for the three distribution types and the conversions.

Summary:
`ms` units in milliseconds, converted to seconds for prometheus
`h` no units, no conversion
`d` no units, no conversion

The benchmark tests were broken, so the metrics were added back to them to get the benchmark working. That commit can be removed if necessary.
Before:
```
$ go test -bench=.
goos: darwin
goarch: amd64
pkg: github.com/prometheus/statsd_exporter
BenchmarkUDPListener1-12                   79335             12802 ns/op
BenchmarkUDPListener5-12                    3866            304820 ns/op
BenchmarkUDPListener50-12                     33          32112409 ns/op
BenchmarkExporterListener-12                 183           6583010 ns/op
PASS
ok      github.com/prometheus/statsd_exporter   5.714s

```

After:
```
$ go test -bench=.
goos: darwin
goarch: amd64
pkg: github.com/prometheus/statsd_exporter
BenchmarkUDPListener1-12                   80152             13512 ns/op
BenchmarkUDPListener5-12                    3638            331360 ns/op
BenchmarkUDPListener50-12                     31          34365441 ns/op
BenchmarkExporterListener-12                 183           6347688 ns/op
PASS
ok      github.com/prometheus/statsd_exporter   5.548s
```
 These results show no significant increase in execution time, as the difference is within the variation observed on repeated benchmarking.