@glightfoot 2020-06-12
ObserverEvent makes sense and is nice and consistent. I pushed up that change now. The only thing I can think of now is that there are plenty of stats calls that still reference timer event:
```				
b.EventStats.WithLabelValues("timer").Inc()
```

I think it makes sense to change those to "observer", but that would be another breaking change to metrics usage, so let me know what you'd like to see.