@glightfoot 2020-06-12
The other thing I can see is there are still lots of references to Timer in the Mapper pkg. I can change them, but then that breaks the configs since things like `timer_type` in the defaults would become `observer_type`, which would need some explaining in the documentation