@matthiasr 2020-06-15
Ah, you ask all the good questions. I'm fine with "breaking" the metrics, a note in the changelog will help, which I will add anyway.

I am less willing to break configurations, how hard would it be to change the primary option but maintain compatibility? (i.e. document `observer_type`, deprecate `timer_type` but treat them the same)?