@glightfoot 2020-06-15
Ok I'll change the metrics now. I haven't looked too much at the mapper config code, so I am not sure how simple of a change that will be