@glightfoot 2020-06-15
I added some custom unmarshalers to handle both keys. `observer_type` will always override `timer_type`. Let me know what you think about this approach. Also note that while this passes all test right now, I haven't yet compiled it and run a full test