@glightfoot 2020-06-18
I built this branch and tested with the following:

```
$ cat << EOF > test-config.yaml
defaults:
  timer_type: summary
mappings:
- match: "(.*)"
  name: "$1"
  match_type: regex
EOF

$ ./statsd_exporter --statsd.mapping-config=test-config.yaml &

$ for x in `seq 1 10`; do
    echo "foo_ms:200|ms" | nc -w 1 -cu localhost 9125
    echo "foo_h:200|h" | nc -w 1 -cu localhost 9125
    echo "foo_d:200|d" | nc -w 1 -cu localhost 9125
done

$ curl -s localhost:9102/metrics | grep -E "foo|statsd_exporter_events_total"
# HELP foo_d Metric autogenerated by statsd_exporter.
# TYPE foo_d summary
foo_d{quantile="0.5"} 200
foo_d{quantile="0.9"} 200
foo_d{quantile="0.99"} 200
foo_d_sum 2000
foo_d_count 10
# HELP foo_h Metric autogenerated by statsd_exporter.
# TYPE foo_h summary
foo_h{quantile="0.5"} 200
foo_h{quantile="0.9"} 200
foo_h{quantile="0.99"} 200
foo_h_sum 2000
foo_h_count 10
# HELP foo_ms Metric autogenerated by statsd_exporter.
# TYPE foo_ms summary
foo_ms{quantile="0.5"} 0.2
foo_ms{quantile="0.9"} 0.2
foo_ms{quantile="0.99"} 0.2
foo_ms_sum 1.9999999999999998
foo_ms_count 10
# HELP statsd_exporter_events_total The total number of StatsD events seen.
# TYPE statsd_exporter_events_total counter
statsd_exporter_events_total{type="observer"} 30
```

Is there anything else necessary for this PR?